finished search

+++ CoffeeTeria_project/BackEnd/models/user.py
@@ -6,7 +6,7 @@ class User(BaseModel):
     id:Optional[str]=''
     username:str
     password:str
-    balance:Optional[int]=0
+    balance:Optional[float]=0
     type:str #owner customer
     city_name:str
     wish_list:Optional[list[str]] =[]
